// truyen-tranh/src/configs/types.ts
export interface IFilm {
  _id: string;
  title?: string;
  secondary_title?: string;
  description?: string;
  thumbnail?: string | any;
  poster?: string;
  category: any;
  year_release: string;
  duration?: string;
  total_episode?: string | number;
  // list_episode?:
  //   | {
  //       name: string;
  //       list_link: {
  //         title: string;
  //         link: string;
  //       };
  //     }[]
  //   | string
  //   | number;
  list_episode?: {
    name: string;
    list_link: {
      title: string;
      link: string;
    }[];
  }[];

  outstanding?: string;
  language?: string[];
  country?: string[];
  quality?: string;
  director: string[];
  performer?: string[];
  slug?: string | any;
  isDeleted?: boolean;
  rate?: {
    star_number: number;
    _id?: string;
    updatedAt?: string | Date;
    createdAt?: string | Date;
  }[];

  views?: number;
  code: string;
  tags?: string[];
  download_link?: any;
  updatedAt?: string | Date;
  createdAt?: string | Date;
}